% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{editops_apply_vec}
\alias{editops_apply_vec}
\title{Apply Edit Operations to Vector}
\usage{
editops_apply_vec(editops, s1, s2)
}
\arguments{
\item{editops}{A data frame of edit operations (type, src_pos, dest_pos).}

\item{s1}{The source string.}

\item{s2}{The target string.}
}
\value{
A character vector representing the transformed string.
}
\description{
Applies edit operations to transform a string.
}
