% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fuzz_ratio}
\alias{fuzz_ratio}
\title{Simple Ratio Calculation}
\usage{
fuzz_ratio(s1, s2, score_cutoff = 0)
}
\arguments{
\item{s1}{First string.}

\item{s2}{Second string.}

\item{score_cutoff}{Optional score cutoff threshold (default: 0.0).}
}
\value{
A double representing the ratio between the two strings.
}
\description{
Calculates a simple ratio between two strings.
}
\examples{
fuzz_ratio("this is a test", "this is a test!")
}
