% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fuzz_token_ratio}
\alias{fuzz_token_ratio}
\title{Combined Token Ratio}
\usage{
fuzz_token_ratio(s1, s2, score_cutoff = 0)
}
\arguments{
\item{s1}{First string.}

\item{s2}{Second string.}

\item{score_cutoff}{Optional score cutoff threshold (default: 0.0).}
}
\value{
A double representing the combined token ratio between the two strings.
}
\description{
Calculates the maximum ratio of token set ratio and token sort ratio.
}
\examples{
fuzz_token_ratio("fuzzy wuzzy was a bear", "wuzzy fuzzy was a bear")
}
