% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fuzz_token_set_ratio}
\alias{fuzz_token_set_ratio}
\title{Token Set Ratio Calculation}
\usage{
fuzz_token_set_ratio(s1, s2, score_cutoff = 0)
}
\arguments{
\item{s1}{First string.}

\item{s2}{Second string.}

\item{score_cutoff}{Optional score cutoff threshold (default: 0.0).}
}
\value{
A double representing the token set ratio between the two strings.
}
\description{
Compares the unique and common words in the strings and calculates the ratio.
}
\examples{
fuzz_token_set_ratio("fuzzy wuzzy was a bear", "fuzzy fuzzy was a bear")
}
