% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fuzz_token_sort_ratio}
\alias{fuzz_token_sort_ratio}
\title{Token Sort Ratio Calculation}
\usage{
fuzz_token_sort_ratio(s1, s2, score_cutoff = 0)
}
\arguments{
\item{s1}{First string.}

\item{s2}{Second string.}

\item{score_cutoff}{Optional score cutoff threshold (default: 0.0).}
}
\value{
A double representing the token sort ratio between the two strings.
}
\description{
Sorts the words in the strings and calculates the ratio between them.
}
\examples{
fuzz_token_sort_ratio("fuzzy wuzzy was a bear", "wuzzy fuzzy was a bear")
}
