% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_editops}
\alias{get_editops}
\title{Get Edit Operations}
\usage{
get_editops(s1, s2)
}
\arguments{
\item{s1}{The source string.}

\item{s2}{The target string.}
}
\value{
A DataFrame with edit operations.
}
\description{
Generates edit operations between two strings.
}
