% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Hamming Distance}
\usage{
hamming_distance(s1, s2, pad = TRUE)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}

\item{pad}{If true, the strings are padded to the same length (default: TRUE).}
}
\value{
An integer representing the Hamming distance.
}
\description{
Calculates the Hamming distance between two strings.
}
\examples{
hamming_distance("karolin", "kathrin")
}
