% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming_similarity}
\alias{hamming_similarity}
\title{Hamming Similarity}
\usage{
hamming_similarity(s1, s2, pad = TRUE)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}

\item{pad}{If true, the strings are padded to the same length (default: TRUE).}
}
\value{
An integer representing the similarity.
}
\description{
Measures the similarity between two strings using the Hamming metric.
}
\examples{
hamming_similarity("karolin", "kathrin")
}
