% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indel_distance}
\alias{indel_distance}
\title{Indel Distance}
\usage{
indel_distance(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A numeric value representing the Indel distance.
}
\description{
Calculates the insertion/deletion (Indel) distance between two strings.
}
\examples{
indel_distance("kitten", "sitting")
}
