% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indel_normalized_distance}
\alias{indel_normalized_distance}
\title{Normalized Indel Distance}
\usage{
indel_normalized_distance(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A numeric value between 0 and 1 representing the normalized Indel distance.
}
\description{
Calculates the normalized insertion/deletion (Indel) distance between two strings.
}
\examples{
indel_normalized_distance("kitten", "sitting")
}
