% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indel_similarity}
\alias{indel_similarity}
\title{Indel Similarity}
\usage{
indel_similarity(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A numeric value representing the Indel similarity.
}
\description{
Calculates the insertion/deletion (Indel) similarity between two strings.
}
\examples{
indel_similarity("kitten", "sitting")
}
