% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaro_normalized_distance}
\alias{jaro_normalized_distance}
\title{Normalized Jaro Distance}
\usage{
jaro_normalized_distance(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A numeric value representing the normalized Jaro distance.
}
\description{
Calculates the normalized Jaro distance between two strings, a value between 0 and 1.
}
\examples{
jaro_normalized_distance("kitten", "sitting")
}
