% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaro_winkler_distance}
\alias{jaro_winkler_distance}
\title{Jaro-Winkler Distance}
\usage{
jaro_winkler_distance(s1, s2, prefix_weight = 0.1)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}

\item{prefix_weight}{The weight applied to the prefix (default: 0.1).}
}
\value{
A numeric value representing the Jaro-Winkler distance.
}
\description{
Calculates the Jaro-Winkler distance between two strings.
}
\examples{
jaro_winkler_distance("kitten", "sitting")
}
