% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaro_winkler_normalized_similarity}
\alias{jaro_winkler_normalized_similarity}
\title{Similaridade Normalizada Jaro-Winkler}
\usage{
jaro_winkler_normalized_similarity(s1, s2, prefix_weight = 0.1)
}
\arguments{
\item{s1}{Primeira string.}

\item{s2}{Segunda string.}

\item{prefix_weight}{Peso do prefixo (valor padrão: 0.1).}
}
\value{
Um valor numérico representando a similaridade normalizada Jaro-Winkler.
}
\description{
Calcula a similaridade normalizada Jaro-Winkler entre duas strings.
}
\examples{
jaro_winkler_normalized_similarity("kitten", "sitting")
}
