% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcs_seq_distance}
\alias{lcs_seq_distance}
\title{LCSseq Distance}
\usage{
lcs_seq_distance(s1, s2, score_cutoff = NULL)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}

\item{score_cutoff}{Score threshold to stop calculation. Default is the maximum possible value.}
}
\value{
A numeric value representing the LCSseq distance.
}
\description{
Calculates the LCSseq (Longest Common Subsequence) distance between two strings.
}
\examples{
lcs_seq_distance("kitten", "sitting")
}
