% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcs_seq_editops}
\alias{lcs_seq_editops}
\title{LCSseq Edit Operations}
\usage{
lcs_seq_editops(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A data.frame containing the edit operations (substitutions, insertions, and deletions).
}
\description{
Calculates the edit operations required to transform one string into another.
}
\examples{
lcs_seq_editops("kitten", "sitting")
}
