% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcs_seq_normalized_distance}
\alias{lcs_seq_normalized_distance}
\title{Normalized LCSseq Distance}
\usage{
lcs_seq_normalized_distance(s1, s2, score_cutoff = 1)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}

\item{score_cutoff}{Score threshold to stop calculation. Default is 1.0.}
}
\value{
A numeric value representing the normalized LCSseq distance.
}
\description{
Calculates the normalized LCSseq distance between two strings.
}
\examples{
lcs_seq_normalized_distance("kitten", "sitting")
}
