% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcs_seq_normalized_similarity}
\alias{lcs_seq_normalized_similarity}
\title{Normalized LCSseq Similarity}
\usage{
lcs_seq_normalized_similarity(s1, s2, score_cutoff = 0)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}

\item{score_cutoff}{Score threshold to stop calculation. Default is 0.0.}
}
\value{
A numeric value representing the normalized LCSseq similarity.
}
\description{
Calculates the normalized LCSseq similarity between two strings.
}
\examples{
lcs_seq_normalized_similarity("kitten", "sitting")
}
