% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcs_seq_similarity}
\alias{lcs_seq_similarity}
\title{LCSseq Similarity}
\usage{
lcs_seq_similarity(s1, s2, score_cutoff = 0L)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}

\item{score_cutoff}{Score threshold to stop calculation. Default is 0.}
}
\value{
A numeric value representing the LCSseq similarity.
}
\description{
Calculates the LCSseq similarity between two strings.
}
\examples{
lcs_seq_similarity("kitten", "sitting")
}
