% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{levenshtein_distance}
\alias{levenshtein_distance}
\title{Levenshtein Distance}
\usage{
levenshtein_distance(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A numeric value representing the Levenshtein distance.
}
\description{
Calculates the Levenshtein distance between two strings, which represents the minimum number
of insertions, deletions, and substitutions required to transform one string into the other.
}
\examples{
levenshtein_distance("kitten", "sitting")
}
