% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{levenshtein_normalized_distance}
\alias{levenshtein_normalized_distance}
\title{Normalized Levenshtein Distance}
\usage{
levenshtein_normalized_distance(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A numeric value representing the normalized Levenshtein distance.
}
\description{
The normalized Levenshtein distance is the Levenshtein distance divided by the maximum length of the
compared strings, returning a value between 0 and 1.
}
\examples{
levenshtein_normalized_distance("kitten", "sitting")
}
