% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{levenshtein_normalized_similarity}
\alias{levenshtein_normalized_similarity}
\title{Normalized Levenshtein Similarity}
\usage{
levenshtein_normalized_similarity(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A numeric value representing the normalized Levenshtein similarity.
}
\description{
The normalized Levenshtein similarity returns a value between 0 and 1,
indicating how similar the compared strings are.
}
\examples{
levenshtein_normalized_similarity("kitten", "sitting")
}
