% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{levenshtein_similarity}
\alias{levenshtein_similarity}
\title{Levenshtein Similarity}
\usage{
levenshtein_similarity(s1, s2)
}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}
}
\value{
A numeric value representing the Levenshtein similarity.
}
\description{
Levenshtein similarity measures how similar two strings are, based on the minimum number
of operations required to make them identical.
}
\examples{
levenshtein_similarity("kitten", "sitting")
}
