% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opcodes_apply_str}
\alias{opcodes_apply_str}
\title{Apply Opcodes to String}
\usage{
opcodes_apply_str(opcodes, s1, s2)
}
\arguments{
\item{opcodes}{A data frame of opcode transformations (type, src_begin, src_end, dest_begin, dest_end).}

\item{s1}{The source string.}

\item{s2}{The target string.}
}
\value{
The transformed string.
}
\description{
Applies opcodes to transform a string.
}
