% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{osa_editops}
\alias{osa_editops}
\title{Edit Operations Using OSA}
\usage{
osa_editops(s1, s2)
}
\arguments{
\item{s1}{A string to transform.}

\item{s2}{A target string.}
}
\value{
A data frame with the following columns:
\describe{
  \item{operation}{The type of operation (delete, insert, replace).}
  \item{source_position}{The position in the source string.}
  \item{destination_position}{The position in the target string.}
}
}
\description{
Provides the edit operations required to transform one string into another using the OSA algorithm.
}
\examples{
osa_editops("string1", "string2")
}
