% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{osa_normalized_distance}
\alias{osa_normalized_distance}
\title{Normalized Distance Using OSA}
\usage{
osa_normalized_distance(s1, s2, score_cutoff = 1)
}
\arguments{
\item{s1}{A string to compare.}

\item{s2}{Another string to compare.}

\item{score_cutoff}{A threshold for the normalized distance score (default is 1.0).}
}
\value{
A double representing the normalized distance score.
}
\description{
Calculates the normalized OSA distance between two strings.
}
\examples{
osa_normalized_distance("string1", "string2")
}
