% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{osa_normalized_similarity}
\alias{osa_normalized_similarity}
\title{Normalized Similarity Using OSA}
\usage{
osa_normalized_similarity(s1, s2, score_cutoff = 0)
}
\arguments{
\item{s1}{A string to compare.}

\item{s2}{Another string to compare.}

\item{score_cutoff}{A threshold for the normalized similarity score (default is 0.0).}
}
\value{
A double representing the normalized similarity score.
}
\description{
Calculates the normalized similarity between two strings using the Optimal String Alignment (OSA) algorithm.
}
\examples{
osa_normalized_similarity("string1", "string2")
}
