% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{osa_similarity}
\alias{osa_similarity}
\title{Similarity Using OSA}
\usage{
osa_similarity(s1, s2, score_cutoff = 0L)
}
\arguments{
\item{s1}{A string to compare.}

\item{s2}{Another string to compare.}

\item{score_cutoff}{A threshold for the similarity score (default is 0).}
}
\value{
An integer representing the OSA similarity.
}
\description{
Calculates the OSA similarity between two strings.
}
\examples{
osa_similarity("string1", "string2")
}
