% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{postfix_distance}
\alias{postfix_distance}
\title{Postfix Distance}
\usage{
postfix_distance(s1, s2, score_cutoff = NULL)
}
\arguments{
\item{s1}{A string to compare.}

\item{s2}{Another string to compare.}

\item{score_cutoff}{A threshold for the distance score (default is the maximum possible size_t value).}
}
\value{
An integer representing the postfix distance.
}
\description{
Calculates the distance between the postfixes of two strings.
}
\examples{
postfix_distance("string1", "string2")
}
