% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{postfix_normalized_distance}
\alias{postfix_normalized_distance}
\title{Normalized Postfix Distance}
\usage{
postfix_normalized_distance(s1, s2, score_cutoff = 1)
}
\arguments{
\item{s1}{A string to compare.}

\item{s2}{Another string to compare.}

\item{score_cutoff}{A threshold for the normalized distance score (default is 1.0).}
}
\value{
A double representing the normalized postfix distance.
}
\description{
Calculates the normalized distance between the postfixes of two strings.
}
\examples{
postfix_normalized_distance("string1", "string2")
}
