% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{postfix_similarity}
\alias{postfix_similarity}
\title{Postfix Similarity}
\usage{
postfix_similarity(s1, s2, score_cutoff = 0L)
}
\arguments{
\item{s1}{A string to compare.}

\item{s2}{Another string to compare.}

\item{score_cutoff}{A threshold for the similarity score (default is 0).}
}
\value{
An integer representing the postfix similarity.
}
\description{
Calculates the similarity between the postfixes of two strings.
}
\examples{
postfix_similarity("string1", "string2")
}
