% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaro_winkler_normalized_similarity}
\alias{jaro_winkler_normalized_similarity}
\title{Normalized Jaro-Winkler Similarity}
\arguments{
\item{s1}{The first string.}

\item{s2}{The second string.}

\item{prefix_weight}{The weight applied to the prefix (default: 0.1).}
}
\value{
A numeric value representing the normalized Jaro-Winkler similarity.
}
\description{
Calculates the normalized Jaro-Winkler similarity between two strings.
}
\examples{
jaro_winkler_normalized_similarity("kitten", "sitting")
}
